requirements.txt
appdirs==1.4.3
attrs==18.2.0
beautifulsoup4==4.6.3
black==18.9b0
blinker==1.4
bs4==0.0.1
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cssselect==1.0.3
cycler==0.10.0
dlib==19.16.0
fake-useragent==0.1.11
Flask==1.0.2
Flask-Cors==3.0.7
future==0.17.1
google-api-core==1.4.0
google-auth==1.5.1
google-cloud-vision==0.34.0
googleapis-common-protos==1.5.3
grpcio==1.15.0
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
lxml==4.2.5
MarkupSafe==1.0
matplotlib==3.0.1
numpy==1.15.4
oauthlib==3.0.1
opencv-python==3.4.3.18
parse==1.9.0
Pillow==5.3.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pyee==5.0.0
pymongo==3.7.2
pyparsing==2.3.0
pyppeteer==0.0.25
pyquery==1.4.0
python-dateutil==2.7.5
python-dotenv==0.9.1
python-twitter==3.5
pytz==2018.5
requests==2.21.0
requests-html==0.9.0
requests-oauthlib==1.2.0
rsa==4.0
sentry-sdk==0.6.6
six==1.11.0
toml==0.10.0
tqdm==4.28.1
tracery==0.1.1
urllib3==1.23
w3lib==1.19.0
websockets==7.0
Werkzeug==0.14.1
